Refactor Validator class
